BaTool: KdIO Completely Landed

+++ Tools/Sources/bt_wav_writer.cpp
@@ -140,7 +140,7 @@ void BtWavWriter::readWordList()
 
 void BtWavWriter::writeWavHeader(int len)
 {
-    uint32_t buf_i;
+    uint16_t buf_i;
     uint16_t buf_s; //short
 
     file->write("RIFF", 4); // "RIFF" is the father of wav
